Add Initial MongoDB Persistence; Closes #6
Problem
-------

The back end should persist simulations states but it does not. This is because it
does not know how to connect to any database.

Solution
--------

Implement an initial persistence interface using MongoDB. There is a RESTful
HTTP interface defined in TotalJS which maps the appropriate HTTP requests to
model mutations using the Mongoose Mongo JS library. The rudimentary CRUD
operations are implemented but are far from safe/proper.

![1455507215](https://cloud.githubusercontent.com/assets/135916/13039286/629785d4-d35a-11e5-8221-70b1dd97e3ed.png)

Areas of Impact
---------------
- Persistence and Database
- Simulation Models

How to Test
-----------

- [ ] Install MongoDB and ensure it's listening on default ports on `localhost`
- [ ] Install the latests dependency updates `npm install`
- [ ] Start the back end server `node --harmony app.js`
- [ ] Make a request to [http://localhost:8000/simulations](http://localhost:8000/simulations)
- [ ] Use curl to make POST and DELETE HTTP requests to ensure they modify the
  back end collections.

References
----------
- /cc @dill-wood
- [GH Issue](https://github.com/orbitable/mission-control/issues/6)