Update Configuration for Mongo
Problem
=======

The configuration options for mongo are hard coded into the application. This
makes configuring the service at deploy time difficult, ie it is impossible at
the moment to give the service a private username and password to use with
production data.

Solution
========

Update the configuration process to accept environment variables. Then use this
parameterization to connect to the mongo database. The mongo database has been
configured to use authentication to prevent anonymous connections. This is a
preventative security measure given the service will hold authentication data.

Howto Test
==========

- [ ] Tests should pass
- [ ] Running `npm debug.js` should have no regressions