Use Username for User Collections Resource
Problem
-------

You must give the mongo id for a user when accessing the /users/<id>/simulations
endpoint which is difficult to obtain.

Solution
--------

Use the alphanumeric username to reference the user in that resource url. This
provides for friendlier urls when accessing user collections of simulations.
This would prevent bridge from having to do some kind of lookup form a username
to mongo id.